requirements.txt
customtkinter
pillow
opencv-python
numpy
pandas